Moved clipboard utils to core library, fixes #6760 (#7752)
* Moved clipboard utils to core library, fixes #6760

(cherry picked from commit 4fb686d40f51a6a8315924ddfd1bd5d0c26b891c)

* Fixed compilation warnings in cliprdr_utils

(cherry picked from commit 8b7a086411d2563e7a44b71402d4467b7fe95630)

* Fixed backport API mismatch

* Added C++ guards to header

(cherry picked from commit 7626a2ce6a33c0cc976a69c0b8365d961fabe090)
