+++ include/freerdp/channels/cliprdr.h
@@ -22,6 +22,7 @@
 
 #include <freerdp/api.h>
 #include <freerdp/types.h>
+#include <freerdp/utils/cliprdr_utils.h>
 
 #include <winpr/shell.h>
 
@@ -92,17 +93,6 @@ typedef struct _CLIPRDR_MFPICT CLIPRDR_MFPICT;
 extern "C"
 {
 #endif
-
-	FREERDP_API UINT cliprdr_parse_file_list(const BYTE* format_data, UINT32 format_data_length,
-	                                         FILEDESCRIPTORW** file_descriptor_array,
-	                                         UINT32* file_descriptor_count);
-	FREERDP_API UINT cliprdr_serialize_file_list(const FILEDESCRIPTORW* file_descriptor_array,
-	                                             UINT32 file_descriptor_count, BYTE** format_data,
-	                                             UINT32* format_data_length);
-	FREERDP_API UINT cliprdr_serialize_file_list_ex(UINT32 flags,
-	                                                const FILEDESCRIPTORW* file_descriptor_array,
-	                                                UINT32 file_descriptor_count,
-	                                                BYTE** format_data, UINT32* format_data_length);
 
 #ifdef __cplusplus
 }
